+++ requirements.txt
@@ -1,4 +1,4 @@
 langchain-core==0.1.35
 langchain-community==0.0.28
 google-cloud-firestore==2.16.0
-more_itertools==10.2.0
+more-itertools==10.2.0
